Updates to viz_cache_csvs_to_clipped_geospatial script. Now support shp output.
